--- app/components/EditScreenInfo.tsx
@@ -1,82 +1,0 @@
-import { Text, View } from './Themed';
-import Colours from '../constants/Colours';
-import ExternalLink from './ExternalLink';
-import { IEditScreenInfo } from '../types/generic.types';
-import { MonoText } from './StyledText';
-import React from 'react';
-import { StyleSheet } from 'react-native';
-
-const EditScreenInfo: React.FC<IEditScreenInfo> = (
-  props: IEditScreenInfo
-): JSX.Element => {
-  return (
-    <View>
-      <View style={styles.getStartedContainer}>
-        <Text
-          style={styles.getStartedText}
-          lightColor='rgba(0,0,0,0.8)'
-          darkColor='rgba(255,255,255,0.8)'>
-          Open up the code for this screen:
-        </Text>
-
-        <View
-          style={[styles.codeHighlightContainer, styles.homeScreenFilename]}
-          darkColor='rgba(255,255,255,0.05)'
-          lightColor='rgba(0,0,0,0.05)'>
-          <MonoText>{props.path}</MonoText>
-        </View>
-
-        <Text
-          style={styles.getStartedText}
-          lightColor='rgba(0,0,0,0.8)'
-          darkColor='rgba(255,255,255,0.8)'>
-          Change any of the text, save the file, and your app will automatically
-          update.
-        </Text>
-      </View>
-
-      <View style={styles.helpContainer}>
-        <ExternalLink
-          style={styles.helpLink}
-          href='https://docs.expo.io/get-started/create-a-new-app/#opening-the-app-on-your-phonetablet'>
-          <Text style={styles.helpLinkText} lightColor={Colours.light.text}>
-            Tap here if your app doesn't automatically update after making
-            changes
-          </Text>
-        </ExternalLink>
-      </View>
-    </View>
-  );
-};
-
-const styles = StyleSheet.create({
-  getStartedContainer: {
-    alignItems: 'center',
-    marginHorizontal: 50
-  },
-  homeScreenFilename: {
-    marginVertical: 7
-  },
-  codeHighlightContainer: {
-    borderRadius: 3,
-    paddingHorizontal: 4
-  },
-  getStartedText: {
-    fontSize: 17,
-    lineHeight: 24,
-    textAlign: 'center'
-  },
-  helpContainer: {
-    alignItems: 'center',
-    marginHorizontal: 20,
-    marginTop: 15
-  },
-  helpLink: {
-    paddingVertical: 15
-  },
-  helpLinkText: {
-    textAlign: 'center'
-  }
-});
-
-export default EditScreenInfo;
