+++ app/components/__tests__/StyledText.spec.tsx
@@ -1,4 +1,4 @@
-import { MonoText } from '../StyledText';
+import { MonoText } from '../styles/StyledText';
 import renderer from 'react-test-renderer';
 
 describe('<StyledText />', () => {
