+++ app/components/features/SwipeView.tsx
