+++ app/components/layout/Calendar.tsx
@@ -1,12 +1,12 @@
-import { ICalendar, IDayItem, IMonthItem } from '../types/date.types';
+import { ICalendar, IDayItem, IMonthItem } from '../../types/date.types';
 import { Pressable, StyleSheet } from 'react-native';
-import { ScrollView, Text, View } from '../components/Themed';
+import { ScrollView, Text, View } from '../styles/Themed';
 import { useEffect, useState } from 'react';
-import { AppState } from '../redux/store';
-import DayItem from '../components/DayItem';
-import { MonoText } from '../components/StyledText';
-import MonthItem from '../components/MonthItem';
-import { useAppSelector } from '../redux/hooks';
+import { AppState } from '../../redux/store';
+import DayItem from './DayItem';
+import { MonoText } from '../styles/StyledText';
+import MonthItem from './MonthItem';
+import { useAppSelector } from '../../redux/hooks';
 
 const Calendar: React.FC<ICalendar> = ({
   handleCalendarChange
