+++ app/screens/NotFoundScreen.tsx
@@ -1,5 +1,5 @@
 import { StyleSheet, TouchableOpacity } from 'react-native';
-import { Text, View } from '../components/Themed';
+import { Text, View } from '../components/styles/Themed';
 import { RootStackScreenProps } from '../types/navigation.types';
 
 export default function NotFoundScreen({
